Crate just_asc

Have some fun, draw in your terminal! Or someone else’s.

This library provides a “canvas” for you to draw in, rendered in ASCII. Uses might include loading animations in CLI tools, or dorky projector backdrops at your party.

use std::f64::consts::PI;
use crate::just_asc::Draw;

fn main() {
    just_asc::draw(
        just_asc::DEFAULT_CONFIG,
        |grid: &mut just_asc::Grid, frame: usize| {
            grid.circle(50., 50., 50.);

            grid.with_transform(|mut transform| {
                transform.translate(50., 50.);
                transform.rotate(PI / 20. * frame as f64);
                transform.line(0., 10., 0., -40.);
            });

            grid.with_transform(|mut transform| {
                transform.translate(50., 50.);
                transform.rotate(PI / 240. * frame as f64);
                transform.line(0., 5., 0., -20.);
            });
        },
    );
}

Modules

tilesets

Out-of-the-box tilesets for your GridConfig.

Structs

Grid

A Grid is what you draw on. Rather than concerning yourself with each character and how its position maps to your drawing, a Grid gives you a 100x100 canvas, with (0,0) in the top-left corner.

GridConfig

The configuration used when instantiating a grid.

Transform

Transforms are temporary grid “wrappers” that can be freely rotated and translated (moved left-right-up-down).

Constants

DEFAULT_CONFIG

A “good-enough” config to get started: an ASCII tileset and a 72x36 grid.

Traits

Draw

The Draw trait defines the actual shapes you can add to your canvas. This trait is present on both Grid and Transform.

Functions

draw

Our core animation function.

once

Just draw an image once.