Struct just_asc::Grid

pub struct Grid { /* fields omitted */ }

A Grid is what you draw on. Rather than concerning yourself with each character and how its position maps to your drawing, a Grid gives you a 100x100 canvas, with (0,0) in the top-left corner.

grid.line(0., 0., 100., 100.); // a line from top-left to bottom-right
grid.circle(50., 50., 25.); // a circle dead center

Implementations

impl Grid

pub fn transform(&mut self) -> Transform<'_>

Creates a Transform from a Grid. Transforms provide a nice structure for making weird (temporary) transformations to your grid: just throw it out when you’re done!

pub fn with_transform<F>(&mut self, f: F) where
F: Fn(Transform<'_>),

Takes a closure which receives a Transform. This is a convenience function to keep the scope of a Transform nice and clear.