Crate just_asc

Have some fun, draw in your terminal! Or someone else's.

This library provides a "canvas" for you to draw in, rendered in ASCII. Uses might include loading animations in CLI tools, or dorky projector backdrops at your party.

use std::f64::consts::PI;

let config = just_asc::GridConfig {
    cell_width: 40,
    cell_height: 20,
    tileset: just_asc::tilesets::PURE_ASCII,
    max_framerate: Some(1),
    print_timing: false,
};

just_asc::draw(config, |grid: &mut just_asc::Grid, _frame: usize| {
    grid.with_transform(|mut transform| {
        transform.translate(50., 50.);
        transform.circle(0., 0., 50.);
        transform.rotate(PI / 4.);
        transform.line(-25., 0., 25., 0.);
        transform.rotate(PI / 2.);
        transform.line(-25., 0., 25., 0.);
    });
});

Modules

tilesets
Out-of-the-box tilesets for your GridConfig.

Structs

Grid
A Grid is what you draw on. Rather than acting directly on cells, a Grid gives you a 100x100 canvas, with (0,0) in the top-left corner.

GridConfig
The configuration used when instantiating a grid. A few configurations about "cells" are present: a cell represents one character on the rendered grid.

Transform
Transforms are grid "wrappers" that keep a mutable reference to the main Grid. The trick is that they can be freely rotated and translated (moved left-right-up-down). This makes it much easier to draw groups of things that spin, orbit, or travel.

Traits

Draw
The Draw trait defines the actual shapes you can add to your canvas. This trait is present on both Grid and Transform.

Functions

draw
Our core animation function. This kicks off an unending drawing, taking two arguments:

once
Just draw an image once. Takes two arguments: