Trait just_asc::Draw

pub trait Draw {
    fn line(&mut self, x1: f64, y1: f64, x2: f64, y2: f64);

    fn ellipse(&mut self, x: f64, y: f64, a: f64, b: f64, keel: f64);

    fn circle(&mut self, x: f64, y: f64, r: f64);
}

The Draw trait defines the actual shapes you can add to your canvas. This trait is present on both Grid and Transform.

Required methods

fn line(&mut self, x1: f64, y1: f64, x2: f64, y2: f64)

line draws a line on your Grid! the first two arguments are your starting x and y, the latter two arguments are your ending x and y.

grid.line(0., 0., 100., 100.); // a line from top-left to bottom-right
grid.line(100., 0., 0., 100.); // a line from top-right to bottom-left
grid.line(60., 50., 75., 35.);
grid.line(75., 35., 90., 50.);
grid.line(90., 50., 75., 65.);
grid.line(75., 65., 60., 50.);

"_                         .d"
 '¶b                      d"
    ¶b                  d"
      "b,            .d"
       '"_         .d"
          ¶b     .d"   .d¶_
            "b,.d"   .d"  '¶_
             ]##,   ]#      ]#
           .d" '"b   '¶_  .d"
         .d"      "b,  '¶d"
       ._P         '"_
     .d"              "b,
   .d"                  "_
 .d"                      "b,
d"`                         "_

fn ellipse(&mut self, x: f64, y: f64, a: f64, b: f64, keel: f64)

ellipse draws ellipses. The first two parameters are the position of its center, followed by its x axis length and y axis length. The last parameter is its "keel", i.e. its rotation around the center (in radians). Set to 0 if you don't want to rotate! Or PI, if you're feeling spicy.

grid.ellipse(40., 20., 30., 10., 0.);
grid.ellipse(70., 30., 30., 10., PI / 4.);

       .________,       __,
   ._P""        ""¶__d""` "¶,
   P             .d"¶      .[
   b           .d"  d      d
   '"b__      .P__d"`     d`
       '""""""P"`       _P`
             ]`       _P`
             'b_ .__P"`
               '""

fn circle(&mut self, x: f64, y: f64, r: f64)

circle draws a circle, where the first two parameters are the position of its center, and the last is its radius. You could accomplish this with ellipse, but using circle will give a slight performance boost.

grid.circle(25., 25., 30.);
grid.circle(45., 45., 30.);

_P"         "¶_
`             'b
        ._d"""""b_,
      _P"       [ "¶_
     d`         [   'b
    ]`         d`    '[
b,  ]        .d`      [
 "¶_d     __P"        [
    '#""""`          d`
     'b,           .d`
       "¶__     __P"
          '"""""`

Loading content...